LPD-19661 Moves tests to the proper test class as they are asserting the LocalService
