+++ modules/apps/layout/layout-test/src/testIntegration/java/com/liferay/layout/service/test/LayoutServiceTest.java
@@ -7,34 +7,22 @@ package com.liferay.layout.service.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.layout.test.util.LayoutTestUtil;
-import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.Layout;
-import com.liferay.portal.kernel.model.LayoutPrototype;
-import com.liferay.portal.kernel.model.LayoutTypePortlet;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.security.auth.PrincipalException;
 import com.liferay.portal.kernel.security.permission.PermissionChecker;
 import com.liferay.portal.kernel.security.permission.PermissionCheckerFactoryUtil;
 import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
-import com.liferay.portal.kernel.service.LayoutLocalServiceUtil;
 import com.liferay.portal.kernel.service.LayoutService;
 import com.liferay.portal.kernel.service.RoleLocalService;
-import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
 import com.liferay.portal.kernel.test.util.GroupTestUtil;
-import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.UserTestUtil;
-import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
-import com.liferay.sites.kernel.util.Sites;
-
-import java.util.Locale;
-import java.util.Map;
 
 import org.junit.Assert;
 import org.junit.Before;
@@ -114,85 +102,6 @@ public class LayoutServiceTest {
 			PermissionThreadLocal.setPermissionChecker(
 				originalPermissionChecker);
 		}
-	}
-
-	@Test
-	public void testUpdateFriendlyURLMap() throws Exception {
-		Layout layout = LayoutTestUtil.addTypePortletLayout(_group);
-
-		long userId = layout.getUserId();
-
-		layout.setUserId(-1);
-
-		layout = LayoutLocalServiceUtil.updateLayout(layout);
-
-		Map<Locale, String> friendlyURLMap = layout.getFriendlyURLMap();
-
-		friendlyURLMap.put(
-			LocaleUtil.GERMANY,
-			StringPool.SLASH + RandomTestUtil.randomString());
-
-		ServiceContext serviceContext = new ServiceContext();
-
-		serviceContext.setUserId(userId);
-
-		LayoutLocalServiceUtil.updateLayout(
-			_group.getGroupId(), layout.isPrivateLayout(), layout.getLayoutId(),
-			layout.getParentLayoutId(), layout.getNameMap(),
-			layout.getTitleMap(), layout.getDescriptionMap(),
-			layout.getKeywordsMap(), layout.getRobotsMap(), layout.getType(),
-			layout.isHidden(), friendlyURLMap, layout.getIconImage(), null, 0,
-			0, 0, serviceContext);
-	}
-
-	@Test
-	public void testUpdateLookAndFeel() throws Exception {
-		Layout layout = LayoutTestUtil.addTypePortletLayout(_group);
-
-		layout = LayoutLocalServiceUtil.updateLookAndFeel(
-			_group.getGroupId(), false, layout.getLayoutId(),
-			"test_WAR_testtheme", "01", StringPool.BLANK);
-
-		LayoutTypePortlet layoutTypePortlet =
-			(LayoutTypePortlet)layout.getLayoutType();
-
-		layoutTypePortlet.setLayoutTemplateId(
-			layout.getUserId(), "1_column", false);
-
-		LayoutLocalServiceUtil.updateLayout(layout);
-	}
-
-	@Test
-	public void testUpdateTypeSettings() throws Exception {
-		LayoutPrototype layoutPrototype = LayoutTestUtil.addLayoutPrototype(
-			RandomTestUtil.randomString());
-
-		Layout layout = layoutPrototype.getLayout();
-
-		layout = LayoutLocalServiceUtil.updateLayout(layout);
-
-		ServiceContext serviceContext = new ServiceContext();
-
-		serviceContext.setUserId(layout.getUserId());
-
-		LayoutLocalServiceUtil.updateLayout(
-			layout.getGroupId(), layout.isPrivateLayout(), layout.getLayoutId(),
-			layout.getParentLayoutId(), layout.getNameMap(),
-			layout.getTitleMap(), layout.getDescriptionMap(),
-			layout.getKeywordsMap(), layout.getRobotsMap(), layout.getType(),
-			layout.isHidden(), layout.getFriendlyURLMap(),
-			layout.getIconImage(), null, 0, 0, 0, serviceContext);
-
-		Layout updatedLayout = LayoutLocalServiceUtil.getLayout(
-			layout.getPlid());
-
-		UnicodeProperties typeSettingsUnicodeProperties =
-			updatedLayout.getTypeSettingsProperties();
-
-		Assert.assertFalse(
-			"Updating layout prototype should not add property \"" +
-				Sites.LAYOUT_UPDATEABLE + "\"",
-			typeSettingsUnicodeProperties.containsKey(Sites.LAYOUT_UPDATEABLE));
 	}
 
 	@DeleteAfterTestRun
